requirements.txt
Flask
slack_sdk
python-dotenv
openai==0.28